Add pwm start

+++ Src/periphery/pwm/pwm.hpp
@@ -29,9 +29,9 @@ class PwmPeriphery {
     static Logger _logger;
 public:
     // static int8_t init(PwmPin pin);
-    static void set_duration(const PwmPin pwm_pin, uint32_t duration_us);
+    static void set_duration(const PwmPin pwm_pin, uint16_t duration_us);
     static uint32_t get_duration(PwmPin pwm_pin);
-    static void set_duty_cycle(PwmPin pwm_pin, uint8_t duty_cycle_pct);
+    static void set_duty_cycle(PwmPin pwm_pin, uint16_t duty_cycle_pct);
     static void reset(PwmPin pwm_pin);
 };
